more security audit fixes

+++ src/main.c
@@ -28,7 +28,7 @@ void app_main(void) {
     static apdu_handler handlers[INS_MAX];
 
     // TODO: Consider using static initialization of a const, instead of this
-    for (int i = 0; i < INS_MAX; i++) {
+    for (size_t i = 0; i < INS_MAX; i++) {
         handlers[i] = handle_apdu_error;
     }
     handlers[INS_VERSION] = handle_apdu_version;
